+++ backend/helpers/sample_data.js
@@ -1,0 +1,15 @@
+export const archipelagos = [
+	{ id: 1, name: "Covidian Montréal Consortium" },
+	{ id: 2, name: "Secret Group of Destiny" },
+];
+export const islands = [
+	{ id: 1, name: "Raftel", nativeFruit: "Peaches", archipelago_id: 1 },
+	{ id: 2, name: "Montoya", nativeFruit: "Apples", archipelago_id: 1 },
+	{ id: 3, name: "Syracuse", nativeFruit: "Oranges", archipelago_id: 2 },
+];
+export const islanders = [
+	{ id: 1, name: "Francis", island_id: 1 },
+	{ id: 2, name: "Riki", island_id: 2 },
+	{ id: 3, name: "Corrina", island_id: 1 },
+	{ id: 4, name: "Bawb", island_id: 3 },
+];
